CHANGE configuration subtree collapsing and expanding
Make the mechanism of collapsing and expanding data tree more effective.
At the beginning, still only the roots are loaded, but then the whole
subtree is always loaded from the backend. When collapsing, the subtree
is no more deleted, but just hidden so it is not necessary to reload it
when again expanded. When reloading data, only the (at least partially)
visible trees are reloaded, the hidden trees are dropped and subject to
reload on future expansion of such a tree.