backend CHANGE update inventory schemas list after a device connect
Connecting a device may add some new schemas received from the device
into the searchpath location. So far, it was enough to update the
schemas list only when showing schemas inventory, but with linking
schemas (displaying in YANG Explorer) from the configuration data,
it is necessary to have the schemas list updated immediately after the
device is connected.